LayoutTests/fast/css/flex-shrink-calculated-value.html

<!DOCTYPE html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<div id="testDiv" style="position: absolute;"></div>
<script>
description("Tests assigning a calculated value to flex-shrink CSS property.");

var testDiv = document.getElementById("testDiv");

shouldBeEmptyString("testDiv.style['flex-shrink']");
evalAndLog("testDiv.style['flex-shrink'] = 'calc(2 * 3)'");
shouldBeEqualToString("testDiv.style['flex-shrink']", "6");
shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('flex-shrink')", "6");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
