LayoutTests/fast/css/flex-shrink-calculated-value-expected.txt

Tests assigning a calculated value to flex-shrink CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['flex-shrink'] is ""
testDiv.style['flex-shrink'] = 'calc(2 * 3)'
PASS testDiv.style['flex-shrink'] is "6"
PASS window.getComputedStyle(testDiv).getPropertyValue('flex-shrink') is "6"
PASS successfullyParsed is true

TEST COMPLETE
